Upgrade payara version to 5
To avoid an issue with disappearing hazelcast nodes, hopefully.
With payara 4, a node in the hazelcast cluster which disappears triggers
a socket exception because the connection is refused when the
application server tries to send requests to the address where the node
was. This blocks all the worker threads in the server so that all the
scheduled logic stops.